User class added.
The user class as a model has been added to the project and a simple
constructor with all the parameters has been codified as well as
some unit tests for the class and the authentication method.

The class is persistent thanks to the use of Ebean (extending the Model)
and the JPA anotations to set up the email as the identifier.