Task class added.
The model for the task has been added as a class in the models
package. Methods for creating a task and finding those which are
not done yet and are involved to a particular user have been
codified too.

Unit tests for those methods were included in the ModelsTest
class continuing the trend.

Database evolutions for this change and the previous one have
been applied and verified.