Login screen, form and session.
A login screen view has been implemented so that the route /login
redirects to it. A simple form accepting an email and a password
has been added as well on the login.scala.html view. In addition,
that form has been validated and a simple authentication procedure
is followed to log in with an existing user preloaded with our
YAML initial data.