Showing the current user.
The request() method has been used to obtain the username of the
currently logged in user. Now the dashboard shows the projects and
tasks involving that user and the user name and email is shown on
top of the dashboard next to the logout button; that user has been
retrieved from the database using the username of the session and
passing it as a parameter to the views.